refactor: 都道府県名と prefCode をまとめた型定義が人口データについての型定義ファイル内で定義されていたため、都道府県に関する型定義ファイルに分け、import 文を修正
